Split application_form_test.rb into one file per field type
application_form_test.rb had grown to 1552 lines covering every
ApplicationForm field helper (text, textarea, checkbox, select,
password, hidden, number, date, file, radio, autocompleter, field
proxy, upload fields, submit) in one file, making it hard to find
where a given field helper's coverage lives.

Split into test/components/application_form/<type>_field_test.rb,
one per helper, mirroring app/components/application_form/*.rb.
Cross-cutting behavior that isn't specific to one field type (Turbo
form attrs, link_to/class_names availability, FieldWithHelp/
FieldLabelRow's shared help/between rendering, derive_form_id) stays
in application_form_test.rb as the base/core file.

Two existing files needed a merge rather than a fresh split:
- test/components/autocompleter_field_test.rb moved into
  application_form/autocompleter_field_test.rb and absorbed the
  autocompleter-helper tests that were living in the mega test file.
- application_form/radio_field_test.rb (previously covering only
  RadioField's per-choice opts extensions) absorbed the radio_field
  helper tests.

Extracted the shared TestFormClass/render_form/render_upload_form/
render_comment_form/render_radio_field rendering helpers into
test/support/application_form_field_test_helpers.rb (autoloaded from
test/support/, same convention as InatStubHelpers) so every split
file includes the same helpers instead of duplicating them.
